Bubble Sort in C#
A generic class named Bubble Sort. This class will have implementation for all the data types that support +, -, / and *. 
Simple implementation of bubble sort is also in the class you can check that as well if you want just basic bubble sort algorithm.
A basic implementation of how to use this is also presented in program.cs file.